move view to internal namespace
