+++ src/DIPS.Xamarin.UI/Controls/Toast/Toast.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using DIPS.Xamarin.UI.Internal.Xaml;
 using Xamarin.Forms;
 
 namespace DIPS.Xamarin.UI.Controls.Toast
