+++ src/DIPS.Xamarin.UI/Internal/Xaml/ToastView.xaml
@@ -2,11 +2,12 @@
 
 <ContentView xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="DIPS.Xamarin.UI.Controls.Toast.ToastView"
+             x:Class="DIPS.Xamarin.UI.Internal.Xaml.ToastView"
              x:Name="toastView"
 
              xmlns:Local="clr-namespace:DIPS.Xamarin.UI.Controls.Toast;assembly=DIPS.Xamarin.UI"
              xmlns:Tag="clr-namespace:DIPS.Xamarin.UI.Controls.Tag;assembly=DIPS.Xamarin.UI"
+             xmlns:Xaml="clr-namespace:DIPS.Xamarin.UI.Internal.Xaml;assembly=DIPS.Xamarin.UI"
 
              Margin="{Binding PositionY, Source={x:Reference toastView}, Converter={Local:PositionToMarginConverter}}"
 
@@ -14,8 +15,8 @@
              HorizontalOptions="CenterAndExpand">
 
     <ContentView.Content>
-        <Tag:Tag x:DataType="{x:Type Local:ToastView}"
-                 BindingContext="{Binding Source={RelativeSource AncestorType={x:Type Local:ToastView}}}"
+        <Tag:Tag x:DataType="{x:Type Xaml:ToastView}"
+                 BindingContext="{Binding Source={RelativeSource AncestorType={x:Type Xaml:ToastView}}}"
 
                  Text="{Binding Text}"
                  FontSize="{Binding FontSize}"
